# Attempt to retrieve video metadata without a frame decode.